requirements.txt
# Azure Functions
azure-functions>=1.17.0

# ADLS / Blob Storage
azure-storage-blob>=12.19.0
azure-storage-file-datalake>=12.14.0
azure-identity>=1.19.0

# Custom Document Parsing (fallback parsers for .md files and error recovery)
mistune>=3.1.0
python-frontmatter>=1.1.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23

# Custom Text Chunking (same as custom path)
langchain-text-splitters>=0.2.0
tiktoken>=0.6.0

# Azure AI Foundry Services (primary parsing + PII)
azure-ai-contentunderstanding>=1.0.0
azure-ai-textanalytics>=5.3.0

# Foundry LLM Embeddings
openai>=1.12.0

# Azure AI Search (push documents)
azure-search-documents>=11.6.0

# Utilities
python-dotenv>=1.0.1
